+++ outputs.tf
@@ -1,10 +1,9 @@
 output "domain" {
   value = var.domain
 
-  description = <<-END
-    Domain
-    https://go.s3d.com/aws-site-group#domain
-    END
+  description = <<-EOT
+    A domain.
+    EOT
 }
 
 output "ec2" {
@@ -14,44 +13,31 @@ output "ec2" {
     ec2_dns  = v.ec2_dns
   }]
 
-  description = <<-END
-    EC2 Instance
-    https://go.s3d.com/aws-site-group#ec2
-    END
+  description = <<-EOT
+    An EC2 Instance.
+    EOT
 }
 
 output "ec2_key_name" {
   value = var.ec2_key_name
 
-  description = <<-END
-    EC2 Key Name
-    https://go.s3d.com/aws-site-group#ec2_key_name
-    END
+  description = <<-EOT
+    A EC2 Key Name.
+    EOT
 }
 
 output "ecr" {
   value = module.ecr
 
-  description = <<-END
-    ECR
-    https://go.s3d.com/aws-site-group#ecr
-    END
-}
-
-output "name_prefix" {
-  value = local.name_prefix
-
-  description = <<-END
-    Name Prefix
-    https://go.s3d.com/aws-site-group#name_prefix
-    END
+  description = <<-EOT
+    The ECR module.
+    EOT
 }
 
 output "tags" {
   value = local.tags
 
-  description = <<-END
-    Tags
-    https://go.s3d.com/aws-site-group#tags
-    END
+  description = <<-EOT
+    The tags.
+    EOT
 }